requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
clip==1.0
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
ftfy==6.2.0
git-filter-repo==2.38.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.2.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
supervision==0.21.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.51
ultralytics-thop==2.0.0
urllib3==2.2.2
wcwidth==0.2.13
certifi==2024.7.4
charset-normalizer==3.3.2
clip==1.0
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
ftfy==6.2.0
git-filter-repo==2.38.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.2.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scikit-learn==1.3.2
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
supervision==0.21.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.51
ultralytics-thop==2.0.0
urllib3==2.2.2
wcwidth==0.2.13
